Final refactor & rename for v4
