+++ PhigrosLibraryCSharp/UtilityExtension.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics.CodeAnalysis;
+using System.Text;
 
 namespace PhigrosLibraryCSharp;
 internal static class UtilityExtension
@@ -14,5 +15,14 @@ internal static class UtilityExtension
 		T[] values = new T[array.Length];
 		array.CopyTo(values, 0);
 		return values;
+	}
+	internal static string ToHex(this byte[] bytes)
+	{
+		StringBuilder sb = new();
+		for (int i = 0; i < bytes.Length; i++)
+		{
+			sb.Append(bytes[i].ToString("x2"));
+		}
+		return sb.ToString();
 	}
 }
